Merger stegene
