+++ src/sider/eu_eøs/saksbehandling/stegMap/artikkel12_1.js
@@ -1,9 +1,9 @@
 import MKV from "../../../../melosyskodeverk";
 import Steg from "../../../../felleskomponenter/stegvelger/stegMotor/steg";
 import { FANE_STATUS, STEG } from "../../../../felleskomponenter/stegvelger";
-import VurderingArtikkel12_1 from "../../stegKomponenter/vurderingArtikkel12_1";
 import { erVilkarOppfylt, hentVilkar } from "../../../../domeneUtils";
 import * as Utils from "../../../../utils";
+import VurderingArtikkel12_x from "../../stegKomponenter/vurderingArtikkel12_x/vurderingArtikkel12_x";
 
 class Artikkel12_1 extends Steg {
   constructor(propsLight, stegPosisjon) {
@@ -35,26 +35,22 @@ class Artikkel12_1 extends Steg {
     ];
     this.id = STEG.ARTIKKEL_12_1;
     this.tittel = propsLight.konvensjonStorbritanniaToggleEnabled ? "Vurdering arbeidstaker" : "Vurdering av 12.1";
-    this.komponent = VurderingArtikkel12_1;
+    this.komponent = VurderingArtikkel12_x;
     this.samleRelevanteData = (_propsLight) => ({
-      artikkel: { kode: MKV.Koder.vilkaar.FO_883_2004_ART12_1, term: "12.1" },
-      begrunnelser: _propsLight.begrunnelser.art12_1_begrunnelser || [],
       redigerbart: _propsLight.generiskStegRedigerbart,
     });
     this.beregnRelevantUI = (_propsLight) => ({
       harAvklaring,
-      visBegrunnelser12: art12_1.oppfylt === false,
-      visBegrunnelser16: art16_1.oppfylt === false,
-      art12_1,
+      art12_x: art12_1,
+      artikkelNavn: "12.1",
       art16_1,
     });
     this.handlers = {
-      bekreftOgFortsett: this._propsLight.tilgjengeligeHandlers.bekreftOgFortsett,
+      bekreftOgFortsett: propsLight.tilgjengeligeHandlers.bekreftOgFortsett,
       tilbake: propsLight.tilgjengeligeHandlers.tilbake,
-      oppdaterData: (felt, verdi) => this._propsLight.tilgjengeligeHandlers.oppdaterStegData(this.id, felt, verdi),
-      slettData: (data) => this._propsLight.tilgjengeligeHandlers.slettStegData(this.id, data),
+      oppdaterData: (felt, verdi) => propsLight.tilgjengeligeHandlers.oppdaterStegData(this.id, felt, verdi),
+      slettData: (data) => propsLight.tilgjengeligeHandlers.slettStegData(this.id, data),
     };
-
     this._status = FANE_STATUS.OK;
   }
 }
